build system rewrite
This also moves the initialisation of the library in tests into a GTest SetUp
function which prevents build failures due to GTest/CMake calling the
executables while building to enumerate the available tests.
Alongside with a minor update to GTest (latest still supporting C++11),
which is now installed with CMake.
After this, "make install" will properly install the library, the headers and the CMakeConfig/Target.cmake files.

Signed-off-by: Tiziano Müller <tiziano.mueller@hpe.com>