add clang-format & pre-commit config, reformat all
The convention to run git-clang-format was not in place from the start, resulting in different formating across the codebase. Introduce a .clang-format which sets the default indent width to 4 (what is used mostly), based on the LLVM style (which has been used by git-clang-format implicitly due to the absence of any config). Add a pre-commit config to reduce manual labour.

Signed-off-by: Tiziano Müller <tiziano.mueller@hpe.com>